renamed packages renamed FileMetadataRepository to StudyDAO added gitignore added deeper verification for management controller
