--- maestro-app/src/main/java/bio/overture/maestro/infra/adapter/inbound/webapi/ErrorDetails.java
@@ -1,14 +1,0 @@
-package bio.overture.maestro.infra.adapter.inbound.webapi;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-import java.util.Date;
-
-@AllArgsConstructor
-@Getter
-class ErrorDetails {
-    private Date timestamp;
-    private String message;
-    private String details;
-}
