--- maestro-app/src/main/resources/file-centric.mapping.json
@@ -1,27 +1,0 @@
-{
-   "file-centric":{
-      "_all":{
-      	"enabled":false
-      },
-      "dynamic_templates":[
-         {
-            "template_1":{
-               "match":"*",
-               "match_mapping_type":"string",
-               "mapping":{
-                  "type":"keyword"
-               }
-            }
-         }
-      ],
-      "date_detection":false,
-      "properties":{
-         "fileCopies":{
-            "type":"nested"
-         },
-         "donors":{
-            "type":"nested"
-         }
-      }
-   }
-}
