+++ maestro-app/src/test/resources/fixtures/ManagementControllerTest/doc0.json
@@ -1,0 +1,64 @@
+{
+  "object_id": "43d29864-6756-51f0-bb2a-ba7cff860778",
+  "access": "controlled",
+  "study": "PEME-CA",
+  "analysis": {
+    "id": "EGAZ00001254368",
+    "type": "sequencingRead",
+    "state": "PUBLISHED",
+    "study": "PEME-CA",
+    "experiment": {
+      "analysisId": "EGAZ00001254368",
+      "aligned": true,
+      "libraryStrategy": "WGS",
+      "info": { }
+    },
+    "info": null
+  },
+  "file": {
+    "name": "29ff9df727803d20834b9997bc17e970.tumor_MDT-AP-0749_merged.mdup.bam",
+    "format": "BAM",
+    "md5sum": "29ff9df727803d20834b9997bc17e970",
+    "size": 88906416144,
+    "last_modified": null,
+    "index_file": {
+      "object_id": "d819f154-2292-56f7-94f3-bbe03fb45bc5",
+      "name": "29ff9df727803d20834b9997bc17e970.tumor_MDT-AP-0749_merged.mdup.bam.bai",
+      "format": "BAI",
+      "md5sum": "e2a83668aa4cfe4f4048e75212d44f19",
+      "size": 8939816
+    }
+  },
+  "repositories": [
+    {
+      "code": "collab",
+      "organization": "ICGC",
+      "name": "collaboratory",
+      "type": "S3",
+      "country": "CA",
+      "base_url": "https://song.cancercollaboratory.org",
+      "data_path": "/oicr.icgc/data",
+      "metadata_path": "/oicr.icgc.meta/metadata/41ba4fb3-9428-50b5-af6c-d779cd59b04d"
+    }
+  ],
+  "donors": [
+    {
+      "donor_id": "DO232978",
+      "donor_submitted_id": "MDT-AP-0749",
+      "primary_site": "<ADD TO SONG>",
+      "specimen": {
+        "id": "SP201301",
+        "type": "Primary tumour - solid tissue",
+        "submitted_id": "MDT-AP-0749_tumor_specimen",
+        "sample": {
+          "id": "SA604924",
+          "submitted_id": "MDT-AP-0749_tumor",
+          "type": "DNA",
+          "info": { }
+        },
+        "info": { }
+      }
+    }
+  ]
+
+}
