+++ maestro-app/src/test/resources/fixtures/ManagementControllerTest/doc1.json
@@ -1,0 +1,65 @@
+{
+  "object_id": "9ce9358d-c93a-5f83-8032-4addcb84b51a",
+  "access": "controlled",
+  "study": "PEME-CA",
+  "analysis": {
+    "id": "EGAZ00001254247",
+    "type": "sequencingRead",
+    "state": "PUBLISHED",
+    "study": "PEME-CA",
+    "experiment": {
+      "analysisId": "EGAZ00001254247",
+      "aligned": true,
+      "libraryStrategy": "WGS",
+      "info": {}
+    },
+    "info": null
+  },
+  "file": {
+    "name": "b2e40ebb719e1754e99be2e752239639.control_MDT-AP-0432_merged.mdup.bam",
+    "format": "BAM",
+    "md5sum": "b2e40ebb719e1754e99be2e752239639",
+    "size": 77675501639,
+    "last_modified": null,
+    "index_file": {
+      "object_id": "6aa8d318-e168-520e-9d1e-cef127ee6b65",
+      "name": "b2e40ebb719e1754e99be2e752239639.control_MDT-AP-0432_merged.mdup.bam.bai",
+      "format": "BAI",
+      "md5sum": "227a2a11d660a8a53a7375c6623034e2",
+      "size": 8982928
+    }
+  },
+  "repositories": [
+    {
+      "code": "collab",
+      "organization": "ICGC",
+      "name": "collaboratory",
+      "type": "S3",
+      "country": "CA",
+      "base_url": "https://song.cancercollaboratory.org",
+      "data_path": "/oicr.icgc/data",
+      "metadata_path": "/oicr.icgc.meta/metadata/464116e4-afc9-5879-b567-6f54513a32dc"
+    }
+  ],
+  "donors": [
+    {
+      "donor_id": "DO232959",
+      "donor_submitted_id": "MDT-AP-0432",
+      "primary_site": "<ADD TO SONG>",
+      "specimen": {
+        "id": "SP200947",
+        "type": "Normal - blood derived",
+        "submitted_id": "MDT-AP-0432_control_specimen",
+        "sample": {
+          "id": "SA604905",
+          "submitted_id": "MDT-AP-0432_control",
+          "type": "DNA",
+          "info": {
+          }
+        },
+        "info": {
+        }
+      }
+    }
+  ]
+}
