+++ maestro-domain/src/main/java/bio/overture/maestro/domain/api/Indexer.java
@@ -2,7 +2,7 @@ package bio.overture.maestro.domain.api;
 
 import bio.overture.maestro.domain.api.message.IndexResult;
 import bio.overture.maestro.domain.api.message.IndexStudyCommand;
-import bio.overture.maestro.domain.port.outbound.FileDocumentIndexingAdapter;
+import bio.overture.maestro.domain.port.outbound.FileCentricIndexAdapter;
 import lombok.NonNull;
 import reactor.core.publisher.Mono;
 
@@ -12,7 +12,7 @@ import reactor.core.publisher.Mono;
 public interface Indexer {
     /**
      * This method will fetch the specified study from the specified repository
-     * and will invoke the index server {@link FileDocumentIndexingAdapter}
+     * and will invoke the index server {@link FileCentricIndexAdapter}
      * adapter to batch index the resulting documents.
      *
      * @param indexStudyCommand contains the arguments needed to index a single study.
