+++ maestro-domain/src/main/java/bio/overture/masestro/test/Fixture.java
@@ -1,31 +1,71 @@
 package bio.overture.masestro.test;
 
+import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.SneakyThrows;
 import lombok.experimental.UtilityClass;
-import lombok.val;
 
 import java.io.File;
+import java.io.IOException;
 import java.util.Optional;
 
 import static bio.overture.maestro.domain.utility.StringUtilities.inputStreamToString;
 
+/**
+ * Helper to load test fixtures from resources.
+ * This class is only for testing purposes not to be used for non test code.
+ */
 @UtilityClass
 public class Fixture {
 
     private final static String BASE_PATH = "fixtures" + File.separator;
     private final static ObjectMapper MAPPER = new ObjectMapper();
 
+
     @SneakyThrows
     public static <T> T loadJsonFixture(Class clazz, String fileName, Class<T> targetClass) {
-        val json = inputStreamToString(
+        return loadJsonFixture(clazz, fileName, targetClass, MAPPER);
+    }
+
+    /**
+     * Use this overload for generics
+     */
+    @SneakyThrows
+    public static <T> T loadJsonFixture(Class clazz, String fileName, TypeReference<T> type) {
+        return loadJsonFixture(clazz, fileName, type, MAPPER);
+    }
+
+    /**
+     * this overload can be used to load json file and convert it to the target class using a custom mapper
+     *
+     * @param clazz will be used to obtain a class loader to get the resources for.
+     * @param fileName the fixture file we want to load
+     * @param targetClass the target java type we want to convert the json to
+     * @param customMapper in case you want to pre configure a mapper (property name case for example)
+     * @param <T> type parameter of the target class
+     *
+     * @return the converted json file as java type
+     *
+     */
+    @SneakyThrows
+    public static <T> T loadJsonFixture(Class clazz, String fileName, Class<T> targetClass, ObjectMapper customMapper) {
+        String json = loadJsonString(clazz, fileName);
+        return customMapper.readValue(json, targetClass);
+    }
+
+    @SneakyThrows
+    public static <T> T loadJsonFixture(Class clazz, String fileName, TypeReference<T> targetClass, ObjectMapper customMapper) {
+        String json = loadJsonString(clazz, fileName);
+        return customMapper.readValue(json, targetClass);
+    }
+
+    public static String loadJsonString(Class clazz, String fileName) throws IOException {
+        return inputStreamToString(
             Optional.ofNullable(clazz.getClassLoader()
                 .getResource(BASE_PATH + clazz.getSimpleName() + File.separator + fileName)
             )
-            .orElseThrow()
-            .openStream()
+                .orElseThrow()
+                .openStream()
         );
-        return MAPPER.readValue(json, targetClass);
     }
-
 }
