+++ maestro-domain/src/test/resources/fixtures/DefaultIndexerTest/filecentricDocument.json
@@ -2,13 +2,11 @@
   {
     "objectId": "43d29864-6756-51f0-bb2a-ba7cff860778",
     "access": "controlled",
-    "study": [
-      "PEME-CA"
-    ],
+    "study": "PEME-CA",
     "analysis": {
-      "analysisId": "EGAZ00001254368",
-      "analysisType": "sequencingRead",
-      "analysisState": "PUBLISHED",
+      "id": "EGAZ00001254368",
+      "type": "sequencingRead",
+      "state": "PUBLISHED",
       "study": "PEME-CA",
       "experiment": {
         "analysisId": "EGAZ00001254368",
@@ -18,54 +16,49 @@
       },
       "info": null
     },
-    "fileCopies": [
+    "file": {
+      "name": "29ff9df727803d20834b9997bc17e970.tumor_MDT-AP-0749_merged.mdup.bam",
+      "format": "BAM",
+      "md5sum": "29ff9df727803d20834b9997bc17e970",
+      "size": 88906416144,
+      "lastModified": null,
+      "indexFile": {
+        "objectId": "d819f154-2292-56f7-94f3-bbe03fb45bc5",
+        "name": "29ff9df727803d20834b9997bc17e970.tumor_MDT-AP-0749_merged.mdup.bam.bai",
+        "format": "BAI",
+        "md5sum": "e2a83668aa4cfe4f4048e75212d44f19",
+        "size": 8939816
+      }
+    },
+    "repositories": [
       {
-        "analysisId": "EGAZ00001254368",
-        "fileId": "43d29864-6756-51f0-bb2a-ba7cff860778",
-        "fileName": "29ff9df727803d20834b9997bc17e970.tumor_MDT-AP-0749_merged.mdup.bam",
-        "fileFormat": "BAM",
-        "fileMd5sum": "29ff9df727803d20834b9997bc17e970",
-        "fileSize": 88906416144,
-        "lastModified": null,
-        "repoCode": "TEST-REPO",
-        "repoOrg": "org",
-        "repoName": "singer",
-        "repoType": "S3",
-        "repoCountry": "CA",
-        "repoBaseUrl": "http://song.sing.sung",
-        "repoDataPath": "/p1/p2/43d29864-6756-51f0-bb2a-ba7cff860778",
-        "repoMetadataPath": "/m1/m2/41ba4fb3-9428-50b5-af6c-d779cd59b04d",
-        "indexFile": {
-          "objectId": "d819f154-2292-56f7-94f3-bbe03fb45bc5",
-          "fileName": "29ff9df727803d20834b9997bc17e970.tumor_MDT-AP-0749_merged.mdup.bam.bai",
-          "fileFormat": "BAI",
-          "fileMd5sum": "e2a83668aa4cfe4f4048e75212d44f19",
-          "repoFileId": null,
-          "fileSize": 8939816
-        }
+        "code": "TEST-REPO",
+        "organization": "org",
+        "name": "singer",
+        "type": "S3",
+        "country": "CA",
+        "baseUrl": "http://song.sing.sung",
+        "dataPath": "/p1/p2",
+        "metadataPath": "/m1/m2/41ba4fb3-9428-50b5-af6c-d779cd59b04d"
       }
     ],
     "donors": [
       {
-        "study": "PEME-CA",
-        "primarySite": "<ADD TO SONG>",
         "donorId": "DO232978",
-        "specimenId": [
-          "SP201301"
-        ],
-        "specimenType": [
-          "Primary tumour - solid tissue"
-        ],
-        "sampleId": [
-          "SA604924"
-        ],
-        "donorSubmitterId": "MDT-AP-0749",
-        "specimenSubmitterId": [
-          "MDT-AP-0749_tumor_specimen"
-        ],
-        "sampleSubmitterId": [
-          "MDT-AP-0749_tumor"
-        ]
+        "donorSubmittedId": "MDT-AP-0749",
+        "primarySite": "<ADD TO SONG>",
+        "specimen": {
+          "id": "SP201301",
+          "type": "Primary tumour - solid tissue",
+          "submittedId": "MDT-AP-0749_tumor_specimen",
+          "sample": {
+            "id": "SA604924",
+            "submittedId": "MDT-AP-0749_tumor",
+            "type": "DNA",
+            "info": {}
+          },
+          "info": {}
+        }
       }
     ]
   }
