+++ run/docker/docker-compose.override.yml
@@ -1,0 +1,11 @@
+version: '3'
+services:
+  maestro:
+    build: ../../maestro-app
+    # this volume allows to change configs without rebuilding image / jar, just need to restart the container
+    volumes:
+      - '../maestro-app/src/main/resources/config:/app/config'
+    environment:
+      - logging.level.root=INFO
+      - logging.level.bio.overture=TRACE
+      - isLocal=true
